Dockerfile to build speaker-recognition easily
This helps build a self contained image containing the speaker-recognition solution that can then be instantiated into a running container in seconds.  Containers can be either persistent or throw-away.  I did this while setting up my own environment containing your software, so I hope this helps you or the users of this project in the future.